requirements.txt
asttokens==3.0.0
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
dotenv==0.9.9
executing==2.2.0
Flask==3.1.1
Flask-JWT-Extended==4.7.1
greenlet==3.2.3
icecream==2.1.5
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
pillow==11.3.0
pycparser==2.22
pycryptodome==3.23.0
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.1.1
reportlab==4.4.2
requests==2.32.4
SQLAlchemy==2.0.41
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
